change/ ファイル・クラス名を変更。一部メソッドのスコープを変更
